requirements.txt
# Dependencias del Bot NutrIA

# Bot de Telegram
pyTelegramBotAPI==4.14.0

# Firebase (base de datos en tiempo real)
firebase-admin==6.2.0

# Groq (IA para análisis de comidas)
groq==0.4.2

# API REST (Flask)
Flask==3.0.0
Flask-CORS==4.0.0

# Funcionalidades principales
python-dotenv==1.0.0
requests==2.31.0
CryptoJS==1.0  # Para hashing de contraseñas

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Utilidades
python-dateutil==2.8.2
